Refactor and remove redundant files
